+++ src/main/java/com/zcpu/tzzmod/action/ActionType.java
@@ -3,7 +3,8 @@ package com.zcpu.tzzmod.action;
 public enum ActionType {
     COMMAND("command"),
     MESSAGE("message"),
-    SOUND("sound");
+    SOUND("sound"),
+    SIGNAL("signal");
 
     private final String id;